Cycles: refactor for readability
